Hours of experiment; one successful JD round-trip!
This comment includes several hours' worth of experiments hidden in the
comments, in case I ever need to review them and decide to take another
course with this implementation.  It also includes, uncommented, a
single successful idea that I finally fleshed out for a JulianDate
class, that is now used by the Topos class to successfully do a single
date conversion.