Add requirements.txt and pin versions
The pip people have decided that versions like "1997" and "2008" are not
production versions of my ephemerides packages.  Which has broken
everything, right in front of a table full of 6 sprinters who are
considering whether to use Python in their future careers.  This commit
explicitly pins the versions to override pip's refusal to install
packages that were installing fine just a few days ago.  May our
posterity be worthy.