Allow earth_tilt() to support float Julian dates
This commit is kind of a big deal: we are returning to full support for
real, honest scalar Julian dates that are a simple float, instead of
really storing them as 1-element arrays and then trying to hide that
fact.  But as this adjustment to JulianDate seems to break nearly
everything, we are doing it over here on a branch.