Confirm that sidereal_time() takes float or array
And to successfully test this, two advances had to be made in the test
suite.  First, a fixture had to be created that returns a series of
JulianDate objects, some of which are scalars and one of which is an
array.  Second, the vcall() function had to be upgraded so that not just
its first argument, but subsequent arguments, can be arrays of values
that should be different with every call.