Make "Astrometric" name a position, not coordinate
Suddenly everything is clear!  Positions are what Skyfield needs to
focus on, not coordinates.  Coordinate systems might not even rank their
own classes; we will see.  Here, we recognize that our GCRS() coordinate
class is nothing of the sort: it is actually the kind of position called
"astrometric"!  The fact that we will (eventually) be able to name that
coordinate in all sorts of coordinate systems is hereby demoted to a
second-class existence as methods, not objects.