Stop carrying true .distance attribute everywhere
Because the NOVAS routines all return true distance, I was blindly
copying a .distance attribute on to every kind of position that Skyfield
generates so that I could use it in tests.  Now, I simply have the test
separately compute the true distance so that I have a value to compare
with the NOVAS return values, instead of mucking up all my objects.