Stop testing vectorization in tests_vs_novas
Before this commit, the tests_vs_novas.py suite was complicated because
its tests of whether our routines match NOVAS were mixed in with tests
of whether a vector of input dates induced our routines to compute a
vector of outputs.  Now, we rely on test_vectorization.py to do the
latter, so the NOVAS tests can be simpler and less indirect.