Add radii to the planet objects offered by the API
Is this a good idea?  The planet objects are really JPL ephemeris data
which does not seem to carry radius information, so it is a bit of a
conceptual confusion to stick a radius on there from another source (the
HORIZONS system).  Plus, how will people react when the planets attached
to raw ephemeris objects lack this information?