Make initial position examples more thorough
Show explicitly how each major kind of position can be generated from a
solar system object, and from a star object, instead of leaving that
implicit or for users to piece together.  Also, switch to a dictionary
list instead of a bullet list and rename the documentation links to
"Read more" since even I myself did not see them for about a minute when
the links were the names of the three kinds of object!