For GitHub #26, fix satellites and documentation
Not only does this make satellite behavior and its documentation
actually agree, and provide a way for satellite calculations to avoid
having to compute the JPL position for the Earth just to determine how a
satellite is positioned relative to a place somewhere else in the local
coordinate system of the Earth, but we also go ahead and expand the
examples at the top of the positions.rst documentation to cover all
basic coordinates someone could want to compute.  You know, just in case
they don't bother reading any farther than the summary at the top.