For GitHub #25 try making Satellites accept arrays
In the process I have determined that Skyfield is something of a mess,
and that in particular its testing regime is convoluted enough that I
cannot add a "print" statement to code that is failing and necessarily
see the result in the failure report, primarily because so much Skyfield
code runs during test generation.  I will try fixing this situation over
this weekend, which should also make the test suite run much faster.