Rename utc_datetime() since it keeps confusing me
This is a classic case of where I want a short pithy name for a
function, but I find that I keep misusing it over and over again.  So,
with due trepidation, I am changing the name to describe its actual
return value.  Will anyone be able to stand using Skyfield now?