Rebuild test_against_novas.py
Why are so many of these floating point numbers different?  My guess is
that I initially generated it on my other laptop, and that something
about the processor architecture is resulting in different behaviors
down at the lowest bits.  Or maybe there was some change to gcc in the
move to Ubuntu 14.10 or somesuch?