Add GAST attribute to JulianDate
This is a satisfying simplification of the code!  The earthlib no longer
needs to import the nutationlib, the sidereal_time() function loses its
second optional argument and becomes simple and single-purpose, and
timelib simply needs to import a second routine from the nutationlib
that it already needed.

And could improve efficiency, depending on how often GAST was needed
repeatedly in a real-world calculation.