Grant 30-day grace period to old leap second files
Users keep seeing a re-download of the leap second file, because the
international authority has not updated, it but let it expire.  This
patch should fix #96 for about a month, hopefully buying us more time to
learn about the actual schedule on which new leap second data file
releases are made.