Explain that ecliptic coordinates are J2000
This provides at least a short-term fix for #106 by making clear that
these methods cannot be used to generate positions relative to the
ecliptic of any given year, but only for J2000.  For the future, it
might be nice to make the methods more capable and to provide
documentation about how to generate ecliptic coordinates for the
position of the ecliptic on other dates.