Start rebuilding atop the new vector classes
The task now is to go back through Skyfield and all its classes' odd
special dunder-attributes that supported operations that we will now be
supporting with vector math, and re-express all the old operations using
sleek new ones that will let the cruft be removed in subsequent diffs.
Exciting!  (Confession: I had to disable a test to get this to pass.)