Added list_bsp method to load
This addresses issue #95. It adds the list_bsp method to load, which queries the remote URLs to figure out what bsp files are available. Just does a simple thing and assumes the touch date is the when the thing was generated and leaves it up to the user to figure out what they should actually use.