Make SGP4 model attributes explicitly accessible
Fix #118 by deciding to not be ashamed about the satellite model hiding
behind the Skyfield satellite object.  While it would also have been
possible to copy all these attributes on to the Skyfield object itself,
that would be more expensive, would be redundant, and would not allow
different satellite models to be swapped in cleanly later without
stranding all the attributes not used in the alternative model.