Move reverse terra routine into earthlib.py
This places it right beneath the routine for which it is the inverse.
While making the move I have, for symmetry with `terra()`, also made the
reverse routine innocent of what a time object looks like; it now just
receives the `gast` as an argument, and no longer takes responsibility
for the conversion to and from the epoch-of-date.

Further work on #42.