Fix when a float is used to initiate a Time object
Previously, when a time object is intiated with a python float, self.tt is a float. However, the repr now expect self.tt to have a length, which would cause an error if self.tt is a python float. This commit passes the tt argument for Time.__init__ through the _to_array function so that self.tt will become a numpy float instead, which has a length.