Allow Empty Time and Angle Objects (#152)
* Fix Time __repr__

The only obstacle to having an empty time was that in a time object the __repr__ couldn't handle when self.tt was an empty array.

* Allow Angles to be Initialized with Lists

Makes it possible to instantiate an Angle object by providing the keywords radians, degrees, or hours with a python list of floats. The Distance and Velocity objects currently already have this capability.

* Allow Empty Angle Objects

Fixes the dstr and hstr methods of Angle objects to still work when self.radians is an empty array.

* Fix Failing Tests, Make String Representations Consistent

Fixed the failing tests, also made the __str__ representation consistent with Distance and Velocity objects. Also made the hstr and dstr for empty angles consistent with the hstr and dstr for non-empty angles.