Separate methods for creating times from JD’s
It was awkward to hide the ability to create times from the Julian day
behind the very last argument in the time constructor argument lists.
Let’s start moving in a better direction that makes the ability more
discoverable, and starts to backtrack on one of those awkward situations
where half the arguments to a given method are always useless.

Fixes #181.