Remove ref_plane from OsculatingElements
Removed this because the OsculatingElements object doesn't use it. The OsculatingElements object produces elements in the same reference plane as the position and velocity vectors that are given to it. I originally added this attribute in case it was necessary to produce ICRF state vectors from an OsculatingElements object, but at the moment this isn't necessary, and it can be added back in if it becomes necessary.