Fix reference_frames keyword in elementslib (#203)
* Removed size and shape attribute

None of the other Skyfield objects have these attributes, so I removed them from OsculatingElements to be consistent

* Remove ref_plane from OsculatingElements

Removed this because the OsculatingElements object doesn't use it. The OsculatingElements object produces elements in the same reference plane as the position and velocity vectors that are given to it. I originally added this attribute in case it was necessary to produce ICRF state vectors from an OsculatingElements object, but at the moment this isn't necessary, and it can be added back in if it becomes necessary.

* Make reference_frame keyword work again

Fixes `osculating_elements_of` so that the reference_frame keyword works. This also makes the tests marked as broken pass.