Revert filename change if download fails
Old behaviour: When an expired file is found it is renamed to 'file.oldX' and a download attempt is made. If this fails the program exits with an OSError.
New behavious: If said download fails the file is renamed again to it's old name. This allows a user to first try "normal" loading and then retry while allowing expired files. This supports systems which are temporarily offline or when data sources are temporarily offline.