After all these years, go back to requirements.txt
Now that `pip` is sometimes able to install packages — if wheels are
available, and the Python version matches the era to which the packages
are pinned — let’s try a switch back to a `requirements.txt`.  Even
though it only works with Python 3.6.  (Python 3.7, which my Ubuntu
laptop features as its default, dies during a `pip install` while trying
to compile AstroPy.)  This might pave the way for easy interoperation
with tools like Binder in the future.