Remove SVG explaining checkout and pull requests
This diagram took effort for someone to write, and has lingered long in
the repository.  But these days, more than 4 years after it was created,
there exist many guides online about checking out source code from
GitHub and getting set up to contribute back to a project with a pull
request.  So for now let’s delete the image from the repository.  If
user questions appear in the future that could be answered using the
diagram, we can always bring the image back.