Improve exception if timescale download fails
Enough users have trouble downloading the official timescale files,
either because of servers that are down or servers that will not accept
connections from their regions, that we should go ahead and let them
know about the workaround right in the message.

Also, let’s go ahead and turn off exception chaining in Python 3 since
users will likely not be helped by the ugly details of urllib.

This fixes #299 because the user will now be given useful advice about
how to proceed in the failure case they encountered.