Switch from path to module name to fix Travis CI
So I wound up breaking the Travis CI run, which until a few commits ago
was a bare inline call `assay --batch skyfield.tests`.  When I switched
it to an invocation of `test-code.sh`, to unify the testing experience
and also to have a central place to put a call to pyflakes, I was
inadvertently also switching to `assay --batch skyfield.tests` which
does *not* work from another directory.  So this should fix it.