Simplify and improve the discrete search tests
Simple stepwise functions are so much easier to specify and reason about
than sine waves, I don’t know why I didn’t think of them at first!  And
the fact that the state is now always advancing to a new integer instead
of bouncing between 1 and 0 means we catch a problem we didn’t before.