Highlight target of URL fragment
Without some highlighting, I get lost when I click on a link to
something in the API docs and the item is down at the bottom of the page
because it happens to be listed last.  Without highlighting to help me,
I always look for the item at the top of the browser window.