Push improved time precision into .observe()
Now that we represent time internally as two floats instead of one, our
light-time correction routine can be improved to generate far less noise
down below the micro-arcsecond level.  That might not sound like much,
but — for example — over the ten seconds around the maximum elongation
of Venus, the elongation varies by only about 5e-5 mas!  Jitter down in
the tens of nano-arcseconds is enough to make a solver think it’s found
several different seconds in which a maximum elongation occurred.