Remove names from VectorFunction init signature
This is a rare backwards compatibility break, risked only because
`VectorFunction` initialization never been documented, and because a
search of GitHub revealed not projects that were instantiating it
directly.  Eliminating the need to define and copy these two strings
from object to object makes the code more tractable, especially as I get
Kepler orbits closer to production status.  The attributes ``center``
and ``target`` themselves will now, when needed, be strings explaining
what celestial bodies are involved.