Give EarthSatellite builtin timescale by default
Earth satellites used to use a timescale without leap seconds to
represent the epoch, which caused problems when users tried using that
date in calculations, so a `ts` parameter was added to the constructor.
But now that Skyfield has the concept of builtin leap second files,
let’s switch Earth satellites to those if no `ts` is passed.