Fix #409 ts.utc() would die if given a date object
The routine would die because `combine()` is a class method, not a
module method; and because it needed a timezone.  A test is here added
to keep the feature fixed.  Also, the timelib module used `dt` for the
`datetime` module but also sometimes as a convenient local name for a
`datetime`, so let’s rename the module to `dt_module` for readability
and consistency.  Inspired by the attempted fix in #410.