As prep for a simplified at(), remove old warning
This four-year-old warning from early 2016 guided users of Skyfield
≤0.6.1 through a breaking change in the API.  But as I now want to
simplify the `at()` / `_at()` system of calls (probably by removing
`_at()` methods entirely), it’s time to retire the warning, in the hope
that none of the remaining ≤0.6.1 installations (of which there might
not be many?) will upgrade without being able to figure out a way around
the breaking change.

This is work towards #433.