Destroy “snag” subsystem: teach Topos to refract
Instead of carrying the elevation of the observer in meters above sea
level around on the `observer_data` data structure, we here make
refraction a responsibility of the `Topos` — which is possible because
we are now carrying the `Topos` itself, rather than its mere string
description, as the `.center` of positions centered on it.  This both
provides a clear API for a planetary topos to plead ignorance of
atmospheric refraction (it can simply opt to not provide the method),
while removing the last use of the awkward and annoying “snag” subsystem
which now disappears with great rejoicing.  Progress towards #433.