Risky: remove “observer_data” parameter from API
This is risky — strictly speaking, I should never change the public API
of any part of Skyfield because it could break user code.

But does there exist even a single user who has seen this parameter
named in the API docs (it’s never mentioned in the main docs nor in code
examples) who decided to read the code and started using it themselves?
My guess is, no, and so I am daring to attempt its deletion.

I plan to revert this if an issue is opened reporting user code that can
detect its absence.