Major pivot of vector str() and repr() logic
The printed representations of vector functions were rather a mess and
had little consistency; most of them, for example, didn’t print both
their center and their target, but just their target.  So let’s now put
the master `VectorFunction` class in charge of how the string ultimately
look, and only ask the subclasses to weigh in on the details.  In
particular, this lets the new ReversedVector properly display itself as
the reverse of whatever the wrapped vector would normally do.