Remove old NamedStar experiment (from version 0.5)
This old experimental `NamedStar` API from 2015, which was never
documented, is now broken because the Hipparcos catalog is not (for the
moment) being downloaded in compressed form (see #454).  Rather than
delay today’s release to fix `NamedStar`, let’s remove it.

This is a dicey and uncharacteristic decision for me.  I usually pride
myself on not breaking anything that appears in a file like `api.py` and
that someone might have started using through their own research of the
code.  But in this case, with the function marked as deprecated for
several years, I am going to chance it.

Thanks to the original author, though, as the experiment led eventually
to the modern approach of loading stars using a Pandas Dataframe!

The actual dictionary of named stars is retained, per promises in #304.