Fix Time[n] for all jd/fraction array combinations
Fixes #458, along with cases where a user might call one of the
Timescale Time-building methods with a single number for JD but an array
for the fraction.  Does NOT fix the case where a Time itself is given a
scalar JD but an array fraction (I don’t think a constructor should be
interfering with values provided by the caller), but does improve the
Time documentation to warn against that case.