Switch to static frames as instances, not classes
Alas!  It was so much fun to have ITRS in all caps, like the acronym is
really written, because it was a class.  But the EOD class looked odd
because to make it readable I needed underscores; and, in the future I
suspect there would have been user confusion about the fact that some
frames were classes while others must be instances.  So let’s make them
all instances for symmetry.  A last snag was my inability yesterday to
get instances printed with docstrings, but the discovery of Sphinx
`autodata` seems to have saved the day.