Untested: try direct introspection of urlopen()
For a side issue mentioned in #372, where a user of Python 2.7.3 got the
error `urlopen() got an unexpected keyword argument 'cafile'`.  Instead
of guessing whether the keyword argument is present in a given Python
version, let’s just directly test for it.  I can’t currently compile
that Python version locally with pyenv, so this is a shot in the dark.