Rename “itrs_velocity”, hoping no one found it
Hopefully no one discovered this public attribute between its
introduction in 1.32 last week and its removal here before the 1.34
release.  It turns out that this vector was not ITRS velocity, which is,
in fact, zero for a Topos; it was instead the almost-GCRS velocity, just
waiting to be rotated into the right position for the date and time.