Stabilize a test in CI
This test broke because of drift between (a) the old `finals2000A.all`
that we keep in the `ci/` directory and that, for the sake of repository
size, we never update; and (b) the `data/iers.npz` file that we keep
updating as new Skyfield versions come out.  So let’s narrow the test
back down to only using the stable file from `ci/`.