Fix #998: detect sunrise/moonrise even at 70°N
First, this makes the main search loop more precise by computing the
actual rate at which the target’s HA is changing, instead of assuming
one rotation per day — which worked okay for the Sun and planets, but
less well for the Moon given its high 13°/day velocity.

Second, this bolts on a final step which, a bit expensively, compares
the final two positions computed in the loop to the actual horizon, and
fits a parabola to the altitude and altitude-rate-of-change to estimate
the exact moment at which the object reaches the horizon.