Test script: die on Python deprecation warnings
A recent issue alerted me to the fact that the test suite was not
detecting Python deprecation warnings.  Turning them on, alas, revealed
that `astropy` and `pandas` were both triggering warnings on their own,
killing the tests.  Here we upgrade them to safe versions, then turn on
deprecation warnings, hopefully permanently.

(Also: so that the test script works in a minimal container, have it
skip running git if git is not installed.)