Don’t pre-build vector functions for eph segments
To reduce the expense of looking up a planet in an ephemeris, each
ephemeris was doing a single scan of its `.segments` in `__init__()` and
caching the resulting vector functions.  But this meant that users could
not usefully edit the `.segments` list — for example, to combine all the
segments from two ephemeris files.  So let’s do everything on-the-fly.